Use size_t for qHash return values
Otherwise I get the following error:
```log
plasma-framework/src/plasma/svg.cpp:39:35: error: non-constant-expression cannot be narrowed from type 'size_t' (aka 'unsigned long') to 'unsigned int' in initializer list [-Wc++11-narrowing]
    std::array<uint, 10> parts = {::qHash(id.width),
                                  ^~~~~~~~~~~~~~~~~
```

+++ autotests/themetest.cpp
@@ -24,7 +24,7 @@
 QString cacheIdHash(const Plasma::SvgPrivate::CacheId &id)
 {
     static const uint seed = 0x9e3779b9;
-    std::array<uint, 10> parts = {::qHash(id.width),
+    std::array<size_t, 10> parts = {::qHash(id.width),
                                   ::qHash(id.height),
                                   ::qHash(id.elementName),
                                   ::qHash(id.filePath),
